requirements.txt
aws-cdk.core==1.121.0

aws-cdk.aws-lambda==1.121.0
aws_cdk.aws_lambda_python==1.121.0
aws_cdk.aws_lambda_event_sources==1.121.0
aws_cdk.aws_sqs==1.121.0
aws_cdk.aws_dynamodb==1.121.0

# テスト
aws-lambda-powertools
